day01-HTML-CSS/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0 
supabase==1.20.0